requirements.txt
streamlit>=1.37.0
spacy==3.5.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
transformers==4.28.1
torch==1.13.1
pandas==1.5.3
plotly==5.15.0
requests==2.31.0
beautifulsoup4==4.12.2
sentencepiece==0.1.98
scikit-learn==1.2.2
numpy==1.24.3
thinc==8.1.12
Cython==0.29.36
aiohttp==3.8.4
streamlit-authenticator==0.4.1
PyYAML==6.0
